Fix: 파일 분리, cd inner shadow 추가
