--- Recorder/CdListView/CdListView/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -1,20 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0.675",
-          "green" : "0.333",
-          "red" : "0.000"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
