--- Recorder/CdListView/CdListView/Assets.xcassets/album/Contents.json
@@ -1,6 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
