--- Recorder/CdListView/CdListView/CdListViewApp.swift
@@ -1,18 +1,0 @@
-//
-//  CdListViewApp.swift
-//  CdListView
-//
-//  Created by 조은비 on 2022/06/11.
-//
-
-import SwiftUI
-
-@main
-struct CdListViewApp: App {
-    
-    var body: some Scene {
-        WindowGroup {
-            HomeView()
-        }
-    }
-}
